updating readme with presentation link
also added a gif of the gltf vscode extension to better demonstrate the features of the ktx2 viewer to the readme. we had the .mp4 video of this gif within the output_images folder already before this commit, but due to file size it couldnt be added to the readme. this commit simply adds a compressed gif version of that video which already existed before the due date into the readme. (TA approved, please dont add late penalty!)